requirements.txt
nextcord==2.5.0
python-dotenv==1.0.0
pymongo~=4.5.0
aiohttp~=3.8.5